refactor: move business logic to methods directory
