+++ functions/sessions/methods/index.js
@@ -1,0 +1,5 @@
+const create = require('./create');
+
+module.exports = {
+  create,
+};
