+++ functions/todos/methods/index.js
@@ -1,0 +1,7 @@
+const create = require('./create');
+const readAll = require('./readAll');
+
+module.exports = {
+  create,
+  readAll,
+};
